backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.1
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
dj-rest-auth==7.0.1
dj-database-url==2.3.0
Django==5.2
django-allauth==65.7.0
django-braces==1.17.0
django-cors-headers==4.9.0
django-haystack==3.3.0
django-model-utils==5.0.0
django-rest-knox==5.0.2
django-star-ratings==0.9.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
idna==3.11
kombu==5.5.3
packaging==25.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
redis==5.2.1
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
swapper==1.4.0
tzdata==2025.2
urllib3==2.6.2
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.8.2
python-dotenv==1.0.1
